Extract Main Menu from Game Loop
