+++ Desarrollo1_Pixeloids/src/main_menu.h
@@ -1,0 +1,22 @@
+#ifndef MAIN_MENU_H
+#define MAIN_MENU_H
+
+namespace pixeloids_luchelli
+{
+
+	void initializeMenuButtons();
+	void updateMenu();
+	void drawMenu();
+	void updateHowToPlay();
+	void drawHowToPlay();
+	void updateCredits();
+	void drawCredits();
+
+} // namespace pixeloids_luchelli
+
+#endif // !MAIN_MENU_H
+
+
+
+
+
